refactor(probability): isolate directed bond reliability contract
