+++ tests/integration/catalog/test_directed_bond_reliability_schema.py
@@ -1,7 +1,7 @@
 """Public schema guidance for directed bond reliability."""
 
 from jacobian.catalog.catalog import Catalog
-from jacobian.math.probability._models import (
+from jacobian.math.probability._directed_bond_reliability import (
     MAX_DIRECTED_BOND_RELIABILITY_ARCS,
     MAX_DIRECTED_BOND_RELIABILITY_DECLARED_VERTICES,
 )
@@ -18,11 +18,11 @@ def test_directed_bond_reliability_schema_exposes_cross_field_contract() -> None
         "Compute directed source-to-target bond connection probability.\n\n"
         f"The request admits at most {MAX_DIRECTED_BOND_RELIABILITY_ARCS} arcs, requires one probability for every\n"
         "arc exactly once (empty for an edgeless graph), and requires distinct\n"
-        "declared source and target vertices.  Traversal work depends only on\n"
-        "relevant vertices (terminals plus arc endpoints), so sparse graphs may\n"
-        "declare very large vertex counts.  It normalizes arc rows\n"
-        "lexicographically, so state indices and the source-bound result do not\n"
-        "depend on input order."
+        "declared source and target vertices. Traversal and replay materialize\n"
+        "only relevant vertices, so sparse graphs may declare very large vertex\n"
+        "counts within the published declared-vertex label bound. It normalizes\n"
+        "arc rows lexicographically, so state indices and the source-bound result\n"
+        "do not depend on input order."
     )
     properties = schema["properties"]
     assert (
